finished wrapper
Wrapper.Broker is finished and can be used in place of the other
broker and bitwise voting and checksums will be applied to
reliably return the correct values.  Use it just as
you would the original broker and it will (using the
original broker) apply the reliability logic.  Note that
remove will not work because IMO remove should never be
used.