Fixed the reliability decrease as copies increase
The problem was when it tried to obtain the candidate id.
If there was no candidate id but the db returned enough values
then it would produce an invalid checksum instead of a keyerror
within wrapper.  The problem stemmed from how the invalid
checksum error was handled.  I changed it so that if there
is an invalid checksum for the candidate id it will make a new
one as it would with a non existing candidate id.
To test the program just run demo-fix.py as you would the other
demo.pys